refactor: Order <-> Table 의존성 분리
