+++ src/main/java/kitchenpos/table/application/response/OrderTableResponse.java
@@ -1,9 +1,9 @@
-package kitchenpos.order.application.response;
+package kitchenpos.table.application.response;
 
 import java.util.List;
 import java.util.stream.Collectors;
 
-import kitchenpos.order.domain.OrderTable;
+import kitchenpos.table.domain.OrderTable;
 
 public class OrderTableResponse {
