+++ src/main/java/kitchenpos/table/application/validator/TableGroupValidatorImpl.java
@@ -1,12 +1,12 @@
-package kitchenpos.order.application.validator;
+package kitchenpos.table.application.validator;
 
 import java.util.List;
 
 import org.springframework.stereotype.Component;
 import org.springframework.util.CollectionUtils;
 
-import kitchenpos.order.domain.OrderTable;
-import kitchenpos.order.domain.TableGroup;
+import kitchenpos.table.domain.OrderTable;
+import kitchenpos.table.domain.TableGroup;
 
 @Component
 public class TableGroupValidatorImpl implements TableGroupValidator {
