+++ src/main/java/kitchenpos/table/dao/repository/JpaOrderTableRepository.java
@@ -1,10 +1,10 @@
-package kitchenpos.order.dao.repository;
+package kitchenpos.table.dao.repository;
 
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import kitchenpos.order.domain.OrderTable;
+import kitchenpos.table.domain.OrderTable;
 
 public interface JpaOrderTableRepository extends JpaRepository<OrderTable, Long> {
