+++ src/main/java/kitchenpos/table/dao/repository/JpaTableGroupRepository.java
@@ -1,8 +1,8 @@
-package kitchenpos.order.dao.repository;
+package kitchenpos.table.dao.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import kitchenpos.order.domain.TableGroup;
+import kitchenpos.table.domain.TableGroup;
 
 public interface JpaTableGroupRepository extends JpaRepository<TableGroup, Long> {
 }
