+++ src/main/java/kitchenpos/table/domain/OrderTable.java
@@ -1,7 +1,5 @@
-package kitchenpos.order.domain;
+package kitchenpos.table.domain;
 
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Objects;
 
 import javax.persistence.Column;
@@ -9,7 +7,6 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.OneToMany;
 import javax.persistence.Table;
 
 /**
@@ -34,9 +31,6 @@ public class OrderTable {
     @Column(name = "empty")
     private boolean empty;
 
-    @OneToMany(mappedBy = "orderTableId")
-    List<Order> orders = new ArrayList<>();
-
     protected OrderTable() {
     }
 
@@ -58,10 +52,6 @@ public class OrderTable {
         }
     }
 
-    public void addOrder(final Order order) {
-        orders.add(order);
-    }
-
     public void group(final Long tableGroupId) {
         changeEmpty(false);
         this.tableGroupId = tableGroupId;
@@ -79,23 +69,10 @@ public class OrderTable {
     }
 
     public void changeEmpty(final boolean empty) {
-        validateCanChangeEmpty();
         if (tableGroupId != null) {
             throw new IllegalArgumentException("could not change empty, table is now grouped");
         }
         this.empty = empty;
-    }
-
-    private void validateCanChangeEmpty() {
-        for (final Order order : orders) {
-            validateOrderCompletion(order);
-        }
-    }
-
-    private void validateOrderCompletion(final Order order) {
-        if (order.isCompletion()) {
-            throw new IllegalArgumentException("can not change to empty, status is not completion");
-        }
     }
 
     public boolean isEmpty() {
