+++ src/test/java/kitchenpos/order/acceptance/OrderAcceptanceTest.java
@@ -13,9 +13,9 @@ import kitchenpos.menu.application.request.MenuProductRequest;
 import kitchenpos.menu.application.request.MenuRequest;
 import kitchenpos.order.application.request.OrderLineItemRequest;
 import kitchenpos.order.application.request.OrderRequest;
-import kitchenpos.order.application.request.OrderTableRequest;
 import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.product.application.request.ProductCreateRequest;
+import kitchenpos.table.application.request.OrderTableRequest;
 
 public class OrderAcceptanceTest extends AcceptanceTest {
